Adição de superclasse e ajustes nos atributos
